[front] Move Val Town MCP server to new metadata pattern
Migrate val_town from the legacy inline tool definition pattern
to the new static metadata pattern with separate files for
metadata, tools, and helpers.

This follows the same structure as zendesk and other recently
migrated servers, improving type safety and code organization.
